requirements.txt
telethon>=1.34.0
openai>=1.12.0
httpx>=0.25.0
pyyaml>=6.0
python-dotenv>=1.0.0